Create CIRV python package
It renames tools to cirv to match the classical python package tree
and adds missing __init__py. It also fixes concrete and abstract
dependencies.

It adds pylint in tox without loading pylintrc which seems too much
permissive. The missing requirements and the pylint fixes will be
managed in next changes.

Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>
Change-Id: Id4e1e290012e53e275f7d64765d388118e7ac235